fix(workspace): address pre-existing clippy/file-size issues from PR #609
- fix(core): replace 'x as f64' with 'f64::from(x)' in cosine_similarity
  (clippy::cast_lossless)
- refactor(redb): extract tests from relationships.rs (530 LOC) into
  relationships_tests.rs (168 LOC) to satisfy <=500 LOC file-size gate
  Uses #[path] attribute to preserve the flat module layout.
